+++ src/models/spot.model.ts
